feat: export Blockchain
