+++ src/blockchain/index.ts
@@ -4,3 +4,4 @@
  */
 
 export { default, IBlockchainApi } from './Blockchain'
+export * as BlockchainUtils from './Blockchain.utils'
